Extract src/diffview.ts — the working-set diff viewer
The overlay that opens when you click the +N −M card: the git_diff call, the
file sections and their collapse handler. 82 lines, identical to HEAD at 65 vs
65. main.ts 2,561 -> 2,499.

Nearly seam-free — one hook, closeFootMenus, because the overlay menus are
mutually exclusive. Everything else it needs is already below it: parsePatch
from ./diff (the one thoroughly-tested piece in the app, so nothing here
re-parses), hunkHtml from ./inspectorview, $ and dropScrim from ./dom.
