+++ src/main.ts
@@ -11,12 +11,11 @@ import { Terminal } from "@xterm/xterm";
 import { FitAddon } from "@xterm/addon-fit";
 import { WebglAddon } from "@xterm/addon-webgl";
 import "@xterm/xterm/css/xterm.css";
-import { parsePatch, type DiffFile } from "./diff";
 import type {
   DiffStat, ExtSession, GitActionResult, Phase, Restorable, Runnable, Sess,
 } from "./types";
 import { isAgent, statusKey, type Engine } from "./types";
-import { $, dropScrim, toast } from "./dom";
+import { $, toast } from "./dom";
 import {
   closeBranchPop, closeWt, openWt, removeWorktreeSession, setWtCloseSession,
   setWtHandToTerminal, setWtLaunch, setWtRenderAll, setWtSetActive,
@@ -31,6 +30,7 @@ import {
 import { forecast5h, forecast7d, rl, setRlLogger, type Forecast } from "./rl";
 import { setTokenScanning, tokenScanning, usageRow } from "./usageview";
 import { closeCafPop, reconcileCaf, setCafHost } from "./caffeinate";
+import { closeDiff, diffOpen, openDiff, setDiffCloseFootMenus } from "./diffview";
 import {
   askTrust, closeRunPicker, closeTaskManager, mgrEdit, openRunPicker,
   openTaskManager, renderMgr, runTargetCtx, setMgrEdit, setTaskUiHost,
@@ -39,7 +39,7 @@ import {
   closeSettings, openSettings, renderSettings, setSettingsHost, setTab, settingsOpen,
 } from "./settings";
 import {
-  dwellText, gaugesHtml, gitBusy, hunkHtml, planHtml, resHtml, RISK_LABEL,
+  dwellText, gaugesHtml, gitBusy, planHtml, resHtml, RISK_LABEL,
   setGitBusy, timelineHtml, verbFor, vitalHtml, wsetHtml,
 } from "./inspectorview";
 import {
@@ -187,6 +187,7 @@ setSettingsHost({ setTheme, effectiveTheme, setSort, setEngine, bumpFont, applyF
 // The task panels run tasks, hand commands to terminals and put panes on stage —
 // none of which they own.
 setCafHost({ closeFootMenus, renderFoot, renderAll });
+setDiffCloseFootMenus(closeFootMenus);
 setTaskUiHost({
   launchTask, openInputPrompt, handToTerminal, activeProjectCtx, activeCwd,
   setActive, renderAll, closePalette,
@@ -1281,64 +1282,7 @@ function submitInputPrompt() {
 // The inspector's HTML builders now live in ./inspectorview; renderInspector
 // below stays here, because painting them into the page is its whole job.
 
-// ---------- working-set diff viewer ----------
-// Clicking the +N −M card opens a read-only peek at the uncommitted diff. The
-// backend (git_diff) hands us one combined unified-diff patch; parsePatch turns it
-// into files/hunks (in ./diff, unit-tested there). Rendering stays here, in the DOM.
-const DSTAT: Record<DiffFile["status"], [string, string]> = {
-  modified: ["M", "s-mod"], added: ["A", "s-add"], deleted: ["D", "s-del"], renamed: ["R", "s-ren"],
-};
-let diffOpen = false;
-// Keyed by folder (workdir/cwd), not session id, so the same viewer serves Episko's
-// own sessions and read-only external ones alike — both are just a git working tree.
-async function openDiff(workdir: string, title: string) {
-  if (!workdir) return;
-  diffOpen = true;
-  $("scrim").classList.add("show");
-  $("diffDlg").classList.add("show");
-  $("diffTitle").textContent = title || basename(workdir);
-  $("diffSub").textContent = "reading working tree…";
-  $("diffBody").innerHTML = `<div class="diff-empty">Reading the working tree…</div>`;
-  try {
-    const res = await invoke<{ patch: string; truncated: boolean } | null>("git_diff", { workdir });
-    if (!diffOpen) return; // closed while the diff was loading
-    renderDiffBody(res ? parsePatch(res.patch) : [], !!res?.truncated);
-  } catch (e) {
-    if (!diffOpen) return;
-    $("diffSub").textContent = "";
-    $("diffBody").innerHTML = `<div class="diff-empty">Couldn't read the diff.<br><span class="mono">${esc(String(e))}</span></div>`;
-  }
-}
-// Several dialogs share the one #scrim, so closing any of them must only drop it
-// once none of the others are still up.
-function closeDiff() {
-  diffOpen = false;
-  $("diffDlg").classList.remove("show");
-  dropScrim();
-}
-function renderDiffBody(files: DiffFile[], truncated: boolean) {
-  const tot = files.reduce((a, f) => ({ add: a.add + f.added, rem: a.rem + f.removed }), { add: 0, rem: 0 });
-  $("diffSub").innerHTML = files.length
-    ? `<span class="add">+${tot.add}</span> <span class="del">−${tot.rem}</span> · ${files.length} file${files.length === 1 ? "" : "s"}`
-    : "";
-  if (!files.length) { $("diffBody").innerHTML = `<div class="diff-empty">No uncommitted changes to show.</div>`; return; }
-  const sections = files.map((f, i) => {
-    const [glyph, cls] = DSTAT[f.status];
-    const name = f.status === "renamed" && f.oldPath
-      ? `<span class="d-old">${esc(f.oldPath)}</span><span class="d-arr">→</span>${esc(f.path)}`
-      : esc(f.path);
-    const counts = f.binary ? `<span class="d-bin">binary</span>`
-      : `<span class="add">+${f.added}</span> <span class="del">−${f.removed}</span>`;
-    const body = f.binary
-      ? `<div class="d-binbody">Binary file — no textual diff.</div>`
-      : f.hunks.map(hunkHtml).join("") || `<div class="d-binbody">No line changes (mode or metadata only).</div>`;
-    return `<div class="dfile" data-fi="${i}">
-      <div class="dfhead" data-dtoggle="${i}"><span class="dchev">▾</span><span class="dstat ${cls}">${glyph}</span><span class="dpath">${name}</span><span class="dcount">${counts}</span></div>
-      <div class="dfbody">${body}</div></div>`;
-  }).join("");
-  const note = truncated ? `<div class="diff-trunc">Diff truncated — too large to show in full. Open a terminal for the complete diff.</div>` : "";
-  $("diffBody").innerHTML = sections + note;
-}
+// The working-set diff viewer moved to ./diffview.
 function renderInspector(s: Sess | null) {
   if (s?.kind === "shell") { renderShellInspector(s); return; }
   if (s?.kind === "task") { renderTaskInspector(s); return; }
@@ -2313,12 +2257,6 @@ $("fShortSeg").addEventListener("click", (e) => { e.stopPropagation(); $("shortP
 $("btnClose").addEventListener("click", () => { if (activeId) closeSession(activeId); });
 
 $("scrim").addEventListener("click", () => { closePalette(); closeWt(); closeDiff(); closeSettings(); closeRunPicker(); closeInputPrompt(); closeTaskManager(); });
-$("diffClose").addEventListener("click", closeDiff);
-// Collapse / expand a file section by clicking its header.
-$("diffBody").addEventListener("click", (e) => {
-  const h = (e.target as HTMLElement).closest<HTMLElement>("[data-dtoggle]");
-  if (h) h.parentElement!.classList.toggle("collapsed");
-});
 $("palInput").addEventListener("input", refreshPal);
 $("palInput").addEventListener("keydown", (e) => {
   const meta = e.metaKey || e.ctrlKey;
